requirements.txt
pyquery==1.4.0
w3lib==1.18.0
requests==2.20.1
Scrapy==1.4.0
pypinyin==0.30.1
demjson==2.2.4
jmespath==0.9.3
kafka-python==1.4.5
redis==3.2.1
PyYAML==3.13